move stories into own dir
